feat: semantic viewer run query button
