+++ packages/frontend/src/features/semanticViewer/components/RunSemanticQueryButton.tsx
@@ -1,6 +1,18 @@
 import { type ResultRow, type SemanticLayerResultRow } from '@lightdash/common';
-import { useEffect, useMemo, type FC } from 'react';
+import {
+    Button,
+    Group,
+    Kbd,
+    MantineProvider,
+    Text,
+    Tooltip,
+} from '@mantine/core';
+import { useOs } from '@mantine/hooks';
+import { IconPlayerPlay } from '@tabler/icons-react';
+import { useCallback, useEffect, useMemo, type FC } from 'react';
+import MantineIcon from '../../../components/common/MantineIcon';
 import { onResults } from '../../../components/DataViz/store/cartesianChartBaseSlice';
+import LimitButton from '../../../components/LimitButton';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { useSemanticLayerQueryResults } from '../api/hooks';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
@@ -8,9 +20,8 @@ import {
     selectAllSelectedFieldNames,
     selectAllSelectedFieldsByKind,
 } from '../store/selectors';
-import { setResults } from '../store/semanticViewerSlice';
+import { setLimit, setResults } from '../store/semanticViewerSlice';
 import { SemanticViewerResultsTransformer } from '../transformers/SemanticViewerResultsTransformer';
-import RunQueryButton from './RunSqlQueryButton';
 
 const mapResultsToTableData = (
     resultRows: SemanticLayerResultRow[],
@@ -32,10 +43,11 @@ const mapResultsToTableData = (
 };
 
 export const RunSemanticQueryButton: FC = () => {
+    const os = useOs();
     const { showToastError } = useToaster();
 
     const allSelectedFields = useAppSelector(selectAllSelectedFieldNames);
-    const { projectUuid, columns, sortBy } = useAppSelector(
+    const { projectUuid, columns, limit, sortBy } = useAppSelector(
         (state) => state.semanticViewer,
     );
     const allSelectedFieldsByKind = useAppSelector(
@@ -100,20 +112,68 @@ export const RunSemanticQueryButton: FC = () => {
         sortBy,
     ]);
 
+    const handleSubmit = useCallback(
+        () =>
+            runSemanticViewerQuery({
+                ...allSelectedFieldsByKind,
+                sortBy,
+            }),
+        [allSelectedFieldsByKind, runSemanticViewerQuery, sortBy],
+    );
+
+    const handleLimitChange = useCallback(
+        (newLimit: number) => dispatch(setLimit(newLimit)),
+        [dispatch],
+    );
+
     return (
-        <RunQueryButton
-            //disabled={selectedTimeDimensions.length === 0}
-            //limit={limit}
-            onSubmit={() =>
-                runSemanticViewerQuery({
-                    ...allSelectedFieldsByKind,
-                    sortBy,
-                })
-            }
-            isLoading={isLoading} /*onLimitChange={(newLimit) => {
-            dispatch(setSqlLimit(newLimit));
-            handleRunQuery(newLimit);
-        }}*/
-        />
+        <Button.Group>
+            <Tooltip
+                label={
+                    <MantineProvider inherit theme={{ colorScheme: 'dark' }}>
+                        <Group spacing="xxs">
+                            <Kbd fw={600}>
+                                {os === 'macos' || os === 'ios' ? '⌘' : 'ctrl'}
+                            </Kbd>
+
+                            <Text fw={600}>+</Text>
+
+                            <Kbd fw={600}>Enter</Kbd>
+                        </Group>
+                    </MantineProvider>
+                }
+                position="bottom"
+                withArrow
+                withinPortal
+                disabled={isLoading}
+            >
+                <Button
+                    size="xs"
+                    pr={limit ? 'xs' : undefined}
+                    leftIcon={<MantineIcon icon={IconPlayerPlay} />}
+                    onClick={handleSubmit}
+                    loading={isLoading}
+                    disabled={allSelectedFields.length === 0}
+                    sx={(theme) => ({
+                        flex: 1,
+                        borderRight: `1px solid ${theme.fn.rgba(
+                            theme.colors.gray[5],
+                            0.6,
+                        )}`,
+                    })}
+                >
+                    {`Run query ${limit ? `(${limit})` : ''}`}
+                </Button>
+            </Tooltip>
+
+            {handleLimitChange !== undefined && (
+                <LimitButton
+                    disabled={allSelectedFields.length === 0}
+                    size="xs"
+                    limit={limit || 500}
+                    onLimitChange={handleLimitChange}
+                />
+            )}
+        </Button.Group>
     );
 };
